refactor out checklist
